\name{odest-package}
\alias{odest-package}
\alias{odest}
\docType{package}
\title{
Reproducing Kernel Hilbert Space based estimation of Systems of Ordinary Differential Equations
}
\description{
These functions implement RKHS based estimation for Systems of Ordinary Differential Equations. They
provide estimate of parameters, confidence intervals and estimates of state variables. 


}
\details{
\tabular{ll}{
Package: \tab oderkhs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-14\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Ivan Vujacic, Javier Gonzalez, Ernst Wit 

Maintainer: Ivan Vujacic <i.vujacic@rug.nl> and Javier González <j.h.gonzalez@sheffield.ac.uk>

}

